# backend/docs/swagger/swagger.json
{
    "basePath": "/api/v1",
    "definitions": {
...
        "ErrorResponse": {
            "properties": {
                "code": {
                    "example": "TEAM_NOT_FOUND",
                    "type": "string"
                },
                "error": {
                    "example": "Team not found",
                    "type": "string"
                },
                "message": {
                    "example": "No team exists with id team-unknown",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "MessageResponse": {
            "properties": {
                "message": {
                    "example": "Hierarchy level deleted successfully",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "TeamsResponse": {
            "properties": {
                "teams": {
                    "items": {
                        "$ref": "#/definitions/AdminTeamDTO"
                    },
                    "type": "array"
                },
                "total": {
                    "example": 6,
                    "type": "integer"
                }
            },
            "type": "object"
        },
...
        "UpdateTeamRequest": {
            "properties": {
                "cadence": {
                    "enum": [
                        "monthly",
                        "quarterly",
                        "half-yearly",
                        "yearly"
                    ],
                    "example": "quarterly",
                    "type": "string"
                },
                "distributionListEmail": {
                    "example": "orion-squad@teams360.demo",
                    "type": "string"
                },
                "name": {
                    "example": "Orion Squad",
                    "type": "string"
                },
                "teamLeadId": {
                    "example": "teamlead1",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
    },
    "info": {
        "contact": {},
        "description": "Team360 Health Check Application REST API. Assesses team health across 11 dimensions with hierarchical org support.",
        "title": "Team360 API",
        "version": "1.0"
    },
    "paths": {
...
        "/admin/teams": {
            "get": {
                "description": "Returns every team in the organization with the full set of admin-facing details: ID, name, team lead ID and name, cadence, distribution list email, member count, and timestamps, along with the total team count. Access is restricted to authenticated users holding the Admin role with level-1 administrative permission.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TeamsResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to query teams",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List all teams (admin)",
                "tags": [
                    "Admin Teams"
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Creates a new team with a name, cadence, and optional team lead and distribution list email. The team ID is auto-generated from the name when the caller does not supply one. When a team lead is set, the team's supervisor chain is automatically derived by walking that lead's reports-to hierarchy. Access is restricted to authenticated users holding the Admin role with level-1 administrative permission.",
                "parameters": [
                    {
                        "description": "Team to create",
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateTeamRequest"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/AdminTeamDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create team",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create a team",
                "tags": [
                    "Admin Teams"
                ]
            }
        },
        "/admin/teams/{id}": {
            "delete": {
                "description": "Permanently deletes the team identified by the path ID. Access is restricted to authenticated users holding the Admin role with level-1 administrative permission.",
                "parameters": [
                    {
                        "description": "Team ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete team",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete a team",
                "tags": [
                    "Admin Teams"
                ]
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "description": "Updates the name, team lead, cadence, and/or distribution list email of the team identified by the path ID; only the fields present in the request body are changed. If the team lead changes, the supervisor chain is re-derived from the new lead's reports-to hierarchy, or cleared entirely if the lead is removed. Returns a 404 if no team exists with that ID. Access is restricted to authenticated users holding the Admin role with level-1 administrative permission.",
                "parameters": [
                    {
                        "description": "Team ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Fields to update",
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateTeamRequest"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AdminTeamDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Team not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update team",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update a team",
                "tags": [
                    "Admin Teams"
                ]
            }
        },
        "/admin/teams/{id}/members": {
            "get": {
                "description": "Returns every member of the team identified by the path ID, each with their user ID, name, and email, plus the total member count. Returns a 404 if no team exists with that ID. Access is restricted to authenticated users holding the Admin role with level-1 administrative permission.",
                "parameters": [
                    {
                        "description": "Team ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TeamMembersResponse"
                        }
                    },
                    "404": {
                        "description": "Team not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to fetch team members",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List a team's members (admin)",
                "tags": [
                    "Admin Teams"
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "description": "Adds the given user as a member of the team identified by the path ID. Returns a 404 if either the team or the user does not exist. Access is restricted to authenticated users holding the Admin role with level-1 administrative permission.",
                "parameters": [
                    {
                        "description": "Team ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "User to add",
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AddTeamMemberRequest"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Team not found or user not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to add team member",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Add a member to a team",
                "tags": [
                    "Admin Teams"
                ]
            }
        },
        "/admin/teams/{id}/members/{userId}": {
            "delete": {
                "description": "Removes the given user's membership from the team identified by the path ID. Returns a 404 if the user is not currently a member of that team. Access is restricted to authenticated users holding the Admin role with level-1 administrative permission.",
                "parameters": [
                    {
                        "description": "Team ID",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "User ID",
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Team member not found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to remove team member",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Remove a member from a team",
                "tags": [
                    "Admin Teams"
                ]
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT token.",
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}